refactor: documents.batchImport -> collections.import
