+++ server/policies/collection.js
@@ -9,6 +9,11 @@ const { allow } = policy;
 
 allow(User, "create", Collection);
 
+allow(User, "import", Collection, (actor) => {
+  if (actor.isAdmin) return true;
+  throw new AdminRequiredError();
+});
+
 allow(User, ["read", "export"], Collection, (user, collection) => {
   if (!collection || user.teamId !== collection.teamId) return false;
